Move AprilTag camera drivers to their own folder
