+++ src/main/java/com/team1701/lib/drivers/cameras/AprilTag/AprilTagCameraIO.java
@@ -1,4 +1,4 @@
-package com.team1701.lib.drivers.cameras;
+package com.team1701.lib.drivers.cameras.AprilTag;
 
 import java.util.ArrayList;
 import java.util.Optional;
