+++ src/main/java/com/team1701/lib/drivers/cameras/AprilTag/CubVisionCameraIO.java
@@ -1,5 +1,6 @@
-package com.team1701.lib.drivers.cameras;
+package com.team1701.lib.drivers.cameras.AprilTag;
 
+import com.team1701.lib.drivers.cameras.AprilTag.CubVisionRawCameraIO.CubVisionCameraInputs;
 import edu.wpi.first.math.geometry.Transform3d;
 import edu.wpi.first.networktables.TimestampedDoubleArray;
 import org.photonvision.simulation.SimCameraProperties;
